LayoutTests/media/wireless-playback-media-player/load-url-mms-fallback.html

<!DOCTYPE html><!-- webkit-test-runner [ WirelessPlaybackMediaPlayerEnabled=true ] -->
<html>
    <head>
        <script src='../media-file.js'></script>
        <script src='../video-test.js'></script>
        <script>
            var route;
            var video;
            var actualURL;

            async function start()
            {
                if (window.internals)
                    internals.mockMediaDeviceRouteController.enabled = true;

                video = document.createElement('video');

                const mediaType = 'video/mp4';
                const mediaURL = findMediaFile('video', '../content/test');
                const mediaSource = new ManagedMediaSource();

                const firstSource = document.createElement('source');
                firstSource.type = mediaType;
                firstSource.src = URL.createObjectURL(mediaSource);
                video.appendChild(firstSource);

                const secondSource = document.createElement('source');
                secondSource.src = mediaURL;
                secondSource.type = mediaType;
                video.appendChild(secondSource);

                await new Promise((resolve) => {
                    mediaSource.addEventListener('sourceopen', resolve, { once: true });
                    document.body.appendChild(video);
                });

                await new Promise((resolve) => {
                    mediaSource.onstartstreaming = async () => {
                        const sourceBuffer = mediaSource.addSourceBuffer(mediaType);
                        const mediaResource = await fetch(mediaURL);
                        const mediaBuffer = await mediaResource.arrayBuffer();
                        sourceBuffer.addEventListener('updateend', resolve, { once: true });
                        sourceBuffer.appendBuffer(mediaBuffer);
                    };
                });

                await waitFor(video, 'canplaythrough');

                run(`video.play()`)
                await waitFor(video, 'playing');

                route = internals.mockMediaDeviceRouteController.createMockMediaDeviceRoute();

                const urlPromise = new Promise((resolve) => {
                    route.setURLCallback(async (url) => {
                        resolve(url);
                    });
                });

                run(`internals.mockMediaDeviceRouteController.activateRoute(route)`);

                const loadedURL = await urlPromise;
                const expectedURL = new URL(mediaURL, document.baseURI).href.substring(document.baseURI.indexOf('/LayoutTests/'));
                actualURL = loadedURL.substring(loadedURL.indexOf('/LayoutTests/'));

                testExpected('actualURL', expectedURL);

                endTest();
            }
        </script>
    </head>
    <body onload='start()'>
        <p>Test that the MMS fallback URL is loaded when a MediaDeviceRoute activates.</p> 
    </body>
</html>
